test(cypress): Make page list trash test retriable
Only insert the attachment once.

When opening the page for the first time
it is in edit mode and the attachment can be inserted.

On test retries the attachment was already there
and the page opened in view mode.
This led to test failures when attempting to switch to view mode.

Signed-off-by: Max <max@nextcloud.com>